+++ packages/web3-validator/CHANGELOG.md
@@ -142,3 +142,5 @@ Documentation:
 -   ESM import bug (#6359)
 
 ## [Unreleased]
+
+-   Added a new function `isUserOperationAllHex` to check if all properties of the `UserOperation` object are valid hexadecimal values.(#6302)